refactor(runtime): extract results collection into agent/turn — #7415
Two functions move to turn/results_collect.rs:
- collect_tool_results: loop-detector feed (Warning/Block append system
  messages, Break bails), media-marker canonicalization, truncation,
  receipts, and the per-call + XML result forms, returned as
  CollectedResults { individual_results, tool_results,
  detection_relevant_output }. The tool_results/individual_results
  accumulator decls fold into it.
- check_identical_output_abort: the time-gated identical-output hash
  abort; the consecutive counter and last-hash stay owned by the loop
  skeleton, passed &mut.
